Regional results extracted
We were able to extract all of the regional meet results.
Next we need to go through all of the season results. This
is more difficult, since each results page is a little different,
and the tables are not labeled, so we would need have the program
look at each page more looking for which race is for which gender,
the distances of each race, etc. to determine what it is looking
for and correctly keep track of all of the team wins and losses.